Refactoring to be able to extract and insert points from the scene
Currently the points are created and deleted every time the user creates
or deletes points and also every time the user uses the "Undo" and "Redo"
actions.
The problem is that when an insert operation is re-done, it always
inserts the point as the last, instead of its original position. Also,
because points are being created and deleted instead of extracted out of
the scene and then inserted back again, when re-doing operations that
involve modifications on the number of points in the scene usually
incorrect pointers are being accessed, making the application crash.